+++ R/loo.R
@@ -741,13 +741,13 @@ subset.psis <- function(x, subset, ...) {
   x
 }
 
-# helper function to compute relative efficiences
-# @param x matrix of posterior draws
+# helper function to compute relative efficiencies
+# @param x matrix of posterior draws (pointwise=FALSE) or function (pointwise=TRUE)
 # @param fit a brmsfit object to extract metadata from
 # @param allow_na allow NA values in the output?
-# @return a numeric vector of length NCOL(x)
+# @return numeric vector of relative efficiencies, length NCOL(x) or nobs for function
 r_eff_helper <- function(x, chain_id, allow_na = TRUE, ...) {
-  if (anyNA(x) || any(is.infinite(x))) {
+  if (!is.function(x) && (anyNA(x) || any(is.infinite(x)))) {
     warning2("Ignoring relative efficiencies due to NA or infinte inputs.")
     return(rep(1, NCOL(x)))
   }